... opencl optical flow ...
	modified:   ../tests/opencl/CLOptflow.cpp
	modified:   ../tests/opencl/CLOptflow.h
	modified:   ../tests/opencl/bilinup.cl
	modified:   ../tests/opencl/calcp1.cl
	modified:   ../tests/opencl/calcp2.cl
	modified:   ../tests/opencl/grad.cl
	modified:   ../tests/opencl/main.cpp
	new file:   ../tests/opencl/main.cppFILE2
	new file:   ../tests/opencl/main.cppFILE3
	modified:   ../tests/opencl/threshold.cl
	modified:   ../tests/opencl/warp.cl
	modified:   ../tests/opencl/warpsub.cl
	modified:   ../tests/optflowwarp/calcflow.cpp

git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@355 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ tests/opencl/warp.cl
@@ -30,7 +30,7 @@ __kernel void WARP( __write_only image2d_t out,  __read_only image2d_t in,  __re
 	alpha = fract( x, &d );
 	idx = ( ( int ) d ) & 0x03;
 	coordin.x = ( float ) ( ( ( int ) d ) >> 2 ) + 0.5f;
-	coordin.y = coord.y + coordin.y + 0.5f;
+	coordin.y = ( float ) coord.y + coordin.y + 0.5f;
 	v = read_imagef( in, sampler, coordin );
 	v1 = index( v, idx );
 	if( idx < 3 )
@@ -44,7 +44,7 @@ __kernel void WARP( __write_only image2d_t out,  __read_only image2d_t in,  __re
 	alpha = fract( x, &d );
 	idx = ( ( int ) d ) & 0x03;
 	coordin.x = ( float ) ( ( ( int ) d ) >> 2 ) + 0.5f;
-	coordin.y = coord.y + coordin.y + 0.5f;
+	coordin.y = ( float ) coord.y + coordin.y + 0.5f;
 	v = read_imagef( in, sampler, coordin );
 	v1 = index( v, idx );
 	if( idx < 3 )
@@ -58,7 +58,7 @@ __kernel void WARP( __write_only image2d_t out,  __read_only image2d_t in,  __re
 	alpha = fract( x, &d );
 	idx = ( ( int ) d ) & 0x03;
 	coordin.x = ( float ) ( ( ( int ) d ) >> 2 ) + 0.5f;
-	coordin.y = coord.y + coordin.y + 0.5f;
+	coordin.y = ( float ) coord.y + coordin.y + 0.5f;
 	v = read_imagef( in, sampler, coordin );
 	v1 = index( v, idx );
 	if( idx < 3 )
@@ -72,7 +72,7 @@ __kernel void WARP( __write_only image2d_t out,  __read_only image2d_t in,  __re
 	alpha = fract( x, &d );
 	idx = ( ( int ) d ) & 0x03;
 	coordin.x = ( ( float ) ( ( ( int ) d ) >> 2 ) ) + 0.5f;
-	coordin.y = coord.y + coordin.y + 0.5f;
+	coordin.y = ( float ) coord.y + coordin.y + 0.5f;
 	v = read_imagef( in, sampler, coordin );
 	v1 = index( v, idx );
 	if( idx < 3 )
